+++ src/enigma/enigmapython/EnigmaDRotorIII.py
@@ -13,6 +13,7 @@ class EnigmaDRotorIII(DynamicTurnoverRotor):
                             position=position, 
                             ring=ring, 
                             notch_indexes=self.notch_indexes,
-                            alphabet=Alphabets.lookup.get('latin_i18n_26chars_lowercase')
+                            alphabet=Alphabets.lookup.get('latin_i18n_26chars_lowercase'),
+                            turnover_function=lambda n, r, l: (n + r) % l
                         )